Split mapping to new file, fix indexing bugs
