+++ propolis/src/hw/virtio/queue.rs
@@ -389,9 +389,8 @@ impl Chain {
         };
         let mut done = 0;
         let total = self.for_remaining_type(true, |addr, len| {
-            let len = usize::min(raw.len(), len);
-            let remain = &mut raw[done..len];
-            if let Some(copied) = mem.read_into(addr, remain) {
+            let remain = &mut raw[done..];
+            if let Some(copied) = mem.read_into(addr, remain, len) {
                 let need_more = copied != remain.len();
 
                 done += copied;
@@ -430,8 +429,8 @@ impl Chain {
         };
         let mut done = 0;
         let total = self.for_remaining_type(false, |addr, len| {
-            let remain = &raw[done..usize::min(raw.len(), done + len)];
-            if let Some(copied) = mem.write_from(addr, remain) {
+            let remain = &raw[done..];
+            if let Some(copied) = mem.write_from(addr, remain, len) {
                 let need_more = copied != remain.len();
 
                 done += copied;
