+++ propolis/src/vmm/mod.rs
@@ -2,6 +2,8 @@
 
 pub mod hdl;
 pub mod machine;
+pub mod mapping;
 
 pub use hdl::*;
 pub use machine::*;
+pub use mapping::*;
